fix: claiming gatt client WriteWithoutResponse crash during Read (#950)
Fixes an issue with `ClaimingGattClientAdapter::WriteWithoutResponse` which caused state of other operations to be reused and overwritten. This could occur when receiving notifications from the peripheral during a pending read, for example, because the notification is acknowledged using a `WriteWithoutResponse`.

fix: unintentional release of claim in `ClaimingGattClientAdapter::WriteWithoutResponse` (#950)
feat: refactor gatt client interface to take handle directly (#950)